refactor: 역할에 맞는 패키지 이동
