+++ src/main/java/Main.java
@@ -1,3 +1,7 @@
+import controller.AttendanceController;
+import view.InputView;
+import view.OutputView;
+
 public class Main {
 
     public static void main(String[] args) {
