+++ src/main/java/domain/AttendanceResult.java
@@ -1,2 +1,4 @@
+package domain;
+
 public record AttendanceResult(Attend attend, AttendStatus attendStatus) {
 }
