+++ src/main/java/domain/AttendanceResults.java
@@ -1,3 +1,5 @@
+package domain;
+
 import java.util.List;
 
 public class AttendanceResults {
