+++ src/main/java/domain/WarningCrew.java
@@ -1,2 +1,4 @@
+package domain;
+
 public record WarningCrew(String name, AttendCount attendCount) {
 }
