+++ src/main/java/domain/WarningStatus.java
@@ -1,3 +1,5 @@
+package domain;
+
 public enum WarningStatus {
     CLEAR, WARNING, INTERVIEW, EXPEL
 }
