+++ src/main/java/util/DateUtil.java
@@ -1,3 +1,6 @@
+package util;
+
+import domain.Attend;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.LocalTime;
