+++ src/test/java/AttendTest.java
@@ -1,6 +1,7 @@
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
+import domain.Attend;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
