+++ src/test/java/AttendsTest.java
@@ -1,10 +1,13 @@
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import domain.Attend;
+import domain.Attends;
 import java.util.ArrayList;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import util.DateUtil;
 
 public class AttendsTest {
